+++ lib/src/floodfill_image.dart
@@ -18,7 +18,7 @@ class FloodFillImage extends StatefulWidget {
   /// List of color that determines to which [Color]
   /// is/are needed to be avoided upon touch.
   /// <br>**Note:** Nearest color shade is applied.
-  final List<Color> avoidColor;
+  final List<Color>? avoidColor;
 
   /// Set fill value [tolerance] that ranges from 0 to 100.
   /// <br>Default value is 8.
@@ -26,31 +26,31 @@ class FloodFillImage extends StatefulWidget {
 
   /// Width of the image.
   /// Parent widget width will be prioritize if it's provided and less than the image width.
-  final int width;
+  final int? width;
 
   /// Height of the image.
   /// Parent widget height will be prioritize if it's provided and less than the image height.
-  final int height;
+  final int? height;
 
   /// Alignment of the image.
-  final AlignmentGeometry alignment;
+  final AlignmentGeometry? alignment;
 
   /// [Widget] to show while the image is being processed on initialization.
   /// <br>It uses [CircularProgressIndicator] by default.
-  final Widget loadingWidget;
+  final Widget? loadingWidget;
 
   /// Callback function that returns the touch position and an [Image] from *dart:ui* when flood fill starts.
   /// <br>**Note:** Touch coordinate is relative to the image dimension.
-  final Function(Offset position,ui.Image image) onFloodFillStart;
+  final Function(Offset position,ui.Image image)? onFloodFillStart;
 
   /// Callback function that returns an [Image] from *dart:ui* when flood fill ended.
-  final Function(ui.Image image) onFloodFillEnd;
+  final Function(ui.Image image)? onFloodFillEnd;
 
   /// Flutter widget that can use paint bucket functionality on the provided image.
   const FloodFillImage(
-      {Key key,
-      @required this.imageProvider,
-      @required this.fillColor,
+      {Key? key,
+      required this.imageProvider,
+      required this.fillColor,
       this.isFillActive = true,
       this.avoidColor,
       this.tolerance = 8,
@@ -60,21 +60,20 @@ class FloodFillImage extends StatefulWidget {
       this.loadingWidget,
       this.onFloodFillStart,
       this.onFloodFillEnd})
-      : assert(imageProvider != null),
-        super(key: key);
+      : super(key: key);
 
   @override
   _FloodFillImageState createState() => _FloodFillImageState();
 }
 
 class _FloodFillImageState extends State<FloodFillImage> {
-  ImageProvider _imageProvider;
-  ImageStream _imageStream;
-  ImageInfo _imageInfo;
-  double _width;
-  double _height;
-  FloodFillPainter _painter;
-  ValueNotifier<Offset> _repainter;
+  ImageProvider? _imageProvider;
+  ImageStream? _imageStream;
+  ImageInfo? _imageInfo;
+  double? _width;
+  double? _height;
+  FloodFillPainter? _painter;
+  ValueNotifier<String>? _repainter;
 
   @override
   void didChangeDependencies() {
@@ -101,23 +100,22 @@ class _FloodFillImageState extends State<FloodFillImage> {
   }
 
   void _getImage() {
-    final ImageStream oldImageStream = _imageStream;
-    _imageStream =
-        _imageProvider.resolve(createLocalImageConfiguration(context));
-    if (_imageStream.key != oldImageStream?.key) {
+    final ImageStream? oldImageStream = _imageStream;
+    _imageStream = _imageProvider?.resolve(createLocalImageConfiguration(context));
+    if (_imageStream?.key != oldImageStream?.key) {
       final ImageStreamListener listener = ImageStreamListener(_updateImage);
       oldImageStream?.removeListener(listener);
-      _imageStream.addListener(listener);
+      _imageStream?.addListener(listener);
     }
   }
 
   void _updateImage(ImageInfo imageInfo, bool synchronousCall) {
     _imageInfo = imageInfo;
-    _width = _imageInfo.image.width.toDouble();
-    _height = _imageInfo.image.height.toDouble();
-    _repainter = ValueNotifier(Offset(-1, -1));
+    _width = _imageInfo?.image.width.toDouble();
+    _height = _imageInfo?.image.height.toDouble();
+    _repainter = ValueNotifier("");
     _painter = FloodFillPainter(
-        image: _imageInfo.image,
+        image: _imageInfo!.image,
         fillColor: widget.fillColor,
         notifier: _repainter,
         onFloodFillStart: widget.onFloodFillStart,
@@ -131,30 +129,30 @@ class _FloodFillImageState extends State<FloodFillImage> {
 
   @override
   void dispose() {
-    _imageStream.removeListener(ImageStreamListener(_updateImage));
+    _imageStream?.removeListener(ImageStreamListener(_updateImage));
     super.dispose();
   }
 
   @override
   Widget build(BuildContext context) {
     if (_painter != null) {
-      _painter.setFillColor(widget.fillColor); //incase we want to update fillColor
-      _painter.setAvoidColor(widget.avoidColor);
-      _painter.setTolerance(widget.tolerance);
-      _painter.setIsFillActive(widget.isFillActive);
+      _painter?.setFillColor(widget.fillColor); //incase we want to update fillColor
+      _painter?.setAvoidColor(widget.avoidColor!);
+      _painter?.setTolerance(widget.tolerance);
+      _painter?.setIsFillActive(widget.isFillActive);
     }
     return (_imageInfo != null)
         ? LayoutBuilder(builder: (context, BoxConstraints constraints) {
-            double w = _width;
-            double h = _height;
+            double w = _width!;
+            double h = _height!;
             if (!constraints.maxWidth.isInfinite) {
-              if (constraints.maxWidth < _width) {
+              if (constraints.maxWidth < _width!) {
                 w = constraints.maxWidth;
               }
             }
 
             if (!constraints.maxHeight.isInfinite) {
-              if (constraints.maxHeight < _height) {
+              if (constraints.maxHeight < _height!) {
                 h = constraints.maxHeight;
               }
             }
@@ -164,15 +162,15 @@ class _FloodFillImageState extends State<FloodFillImage> {
             // print(" w " + w.toString());
             // print(" h " + h.toString());
 
-            _painter.setSize(Size(w, h));
+            _painter!.setSize(Size(w, h));
             return (widget.alignment == null)
-                ? CustomPaint(painter: _painter, size: Size(w, h))
+                ? RepaintBoundary(child: CustomPaint(painter: _painter, size: Size(w, h)))
                 : Align(
-                    alignment: widget.alignment,
+                    alignment: widget.alignment!,
                     child: CustomPaint(painter: _painter, size: Size(w, h)));
           })
         : (widget.loadingWidget == null)
             ? CircularProgressIndicator()
-            : widget.loadingWidget;
+            : widget.loadingWidget!;
   }
 }