+++ lib/src/floodfill_painter.dart
@@ -8,22 +8,23 @@ import 'package:image/image.dart' as img;
 import 'queuelinear_floodfiller.dart';
 
 class FloodFillPainter extends CustomPainter {
-  QueueLinearFloodFiller _filler;
-  double _width;
-  double _height;
-  bool _isFillActive;
-  List<Color> _avoidColor;
+  QueueLinearFloodFiller? _filler;
+  double? _width;
+  double? _height;
+  bool? _isFillActive;
+  List<Color>? _avoidColor;
 
-  ValueNotifier<Offset> notifier;
+  ValueNotifier<String>? notifier;
   ui.Image image;
   Color fillColor;
-  Function(Offset,ui.Image) onFloodFillStart;
-  Function(ui.Image) onFloodFillEnd;
-  Function onInitialize;
+  Function(Offset,ui.Image)? onFloodFillStart;
+  Function(ui.Image)? onFloodFillEnd;
+  Function? onInitialize;
+  Function? onRepainted;
 
   FloodFillPainter(
-      {@required this.image,
-      @required this.fillColor,
+      {required this.image,
+      required this.fillColor,
       this.notifier,
       this.onFloodFillStart,
       this.onFloodFillEnd,
@@ -33,11 +34,11 @@ class FloodFillPainter extends CustomPainter {
   }
 
   void _initFloodFiller() async {
-    ByteData byteData = await image.toByteData(format: ui.ImageByteFormat.png);
+    ByteData byteData = (await image.toByteData(format: ui.ImageByteFormat.png))!;
     var bytes = byteData.buffer.asUint8List();
-    img.Image decoded = img.decodeImage(bytes);
+    img.Image decoded = img.decodeImage(bytes)!;
     _filler = QueueLinearFloodFiller(decoded, img.getColor(fillColor.red, fillColor.green, fillColor.blue, fillColor.alpha));
-    onInitialize();
+    onInitialize!();
   }
 
   void setSize(Size size) {
@@ -47,26 +48,25 @@ class FloodFillPainter extends CustomPainter {
   }
 
   void setFillColor(Color color) {
-    _filler?.setFillColor(
-        img.getColor(color.red, color.green, color.blue, color.alpha));
+    _filler?.setFillColor(img.getColor(color.red, color.green, color.blue, color.alpha));
   }
 
   void setIsFillActive(bool isActive) {
     _isFillActive = isActive;
   }
 
-  void setAvoidColor(List<Color> color) {
+  void setAvoidColor(List<Color>? color) {
     if (color != null) _avoidColor = color;
   }
 
-  void setTolerance(int tolerance) {
+  void setTolerance(int? tolerance) {
     if (tolerance != null) _filler?.setTolerance(tolerance);
   }
 
   bool _checkAvoidColor(int touchColor) {
     if (_avoidColor == null) return false;
 
-    return _avoidColor.any((element) => _isAvoidColor(element, touchColor));
+    return _avoidColor!.any((element) => _isAvoidColor(element, touchColor));
   }
 
   bool _isAvoidColor(Color avoidColor, int touchColor) {
@@ -98,30 +98,29 @@ class FloodFillPainter extends CustomPainter {
 
     if (pX < 0 || pY < 0) return;
 
-    int touchColor = _filler.image.getPixelSafe(pX, pY);
+    int touchColor = _filler!.image!.getPixelSafe(pX, pY);
     if (_checkAvoidColor(touchColor)) return;
+    if (onFloodFillStart != null) onFloodFillStart!(position,image);
 
-    if (onFloodFillStart != null) onFloodFillStart(position,image);
-
-    _filler.setTargetColor(touchColor);
-    await _filler.floodFill(pX, pY);
+    _filler?.setTargetColor(touchColor);
+    await _filler!.floodFill(pX, pY);
 
     ui.decodeImageFromPixels(
-      _filler.image.getBytes(),
-      _filler.image.width,
-      _filler.image.height,
+      _filler!.image!.getBytes(),
+      _filler!.image!.width,
+      _filler!.image!.height,
       ui.PixelFormat.rgba8888,
       (output) async {
         image = output;
-        notifier.value = position;
-        if (onFloodFillEnd != null) onFloodFillEnd(output);
+        notifier!.value = position.toString() + touchColor.toString();
+        if (onFloodFillEnd != null) onFloodFillEnd!(output);
       },
     );
   }
 
   @override
-  bool hitTest(Offset position) {
-    if (_isFillActive) fill(position);
+  bool? hitTest(Offset position) {
+    if (_isFillActive!) fill(position);
     return super.hitTest(position);
   }
 
@@ -151,3 +150,5 @@ class FloodFillPainter extends CustomPainter {
     return true;
   }
 }
+
+